homework07/elevennote/requirements/base.txt
Django==2.0.1
psycopg2==2.7.3.2
python-decouple==3.1
pytz==2017.3
uWSGI==2.0.15
django-widget-tweaks==1.4.1
django-wysiwyg==0.8.0
djangorestframework==3.7.7
djangorestframework-jwt==1.11.0
